+++ src/react-app/states/State5.tsx
@@ -1,0 +1,11 @@
+import RomanticPoster from "./RomanticPoster";
+
+export default function State5() {
+	return (
+		<RomanticPoster
+			text="CHECK OUT MY HOT BOYFRIEND STROKE STROKE VACUUM VACUUM"
+			src="/jonte.jpg"
+			alt="Jonte"
+		/>
+	);
+}